# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VORA API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request an OTP",
        "description": "Returns 202 with an identical body whether or not the phone is registered. Rate limited per phone (3/hour, 10/day), per IP (20/hour), and behind a global circuit breaker that halts sending outright if system-wide volume crosses a threshold.",
        "operationId": "request_otp_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or schema failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Semantically invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Carries Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet (Phase 0 stub)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify an OTP and open a session",
        "description": "Creates the account on first successful verification. The user row is not created at request time, so an unverified number never produces one.",
        "operationId": "verify_otp_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or schema failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Semantically invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Carries Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet (Phase 0 stub)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Rotate the refresh token",
        "description": "The presented token is invalidated by this call. Presenting it again is treated as theft and revokes the entire token family, signing out every device in that chain.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or schema failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Semantically invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Carries Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet (Phase 0 stub)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke the current session",
        "description": "With a refresh_token in the body, revokes that token's family, which signs out this device. With no body, revokes every family for the caller, which signs out all their devices. Either way a family is revoked whole: revoking one link would leave the other devices in that chain signed in, which is not what a user means by log out.\n\nThe body is optional so that §6's bodyless POST /auth/logout keeps working exactly as specified.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LogoutRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request or schema failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role-forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Semantically invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Carries Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented yet (Phase 0 stub)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessibilityProfile": {
        "properties": {
          "requires_ramp": {
            "type": "boolean",
            "title": "Requires Ramp",
            "default": false
          },
          "requires_boot_space": {
            "type": "boolean",
            "title": "Requires Boot Space",
            "default": false
          },
          "requires_front_seat": {
            "type": "boolean",
            "title": "Requires Front Seat",
            "default": false
          },
          "requires_driver_assist": {
            "type": "boolean",
            "title": "Requires Driver Assist",
            "default": false
          },
          "allows_guide_animal": {
            "type": "boolean",
            "title": "Allows Guide Animal",
            "default": false
          },
          "prefers_text_contact": {
            "type": "boolean",
            "title": "Prefers Text Contact",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AccessibilityProfile",
        "description": "Standing vehicle requirements for this user's trips.\n\nRead I9 before adding a field here. Anything that describes the person\nrather than the vehicle they need does not belong in this model, in this\ndatabase, or in this country's legal envelope."
      },
...
      "AuthSessionResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "AuthSessionResponse"
      },
...
      "DriverProfile": {
        "properties": {
          "kyc_status": {
            "$ref": "#/components/schemas/KycStatus"
          },
          "kyc_reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kyc Reviewed At"
          },
          "rating_avg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Avg"
          },
          "is_online": {
            "type": "boolean",
            "title": "Is Online",
            "default": false
          },
          "seats_free": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 0.0,
            "title": "Seats Free",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kyc_status"
        ],
        "title": "DriverProfile",
        "description": "Present on the user object only when role == driver."
      },
      "ErrorDetailModel": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetailModel"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailModel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The §6 envelope. Every non-2xx response in this API has this shape."
      },
...
      "KycStatus": {
        "type": "string",
        "enum": [
          "pending",
          "verified",
          "rejected",
          "suspended"
        ],
        "title": "KycStatus"
      },
...
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096,
                "minLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LogoutRequest",
        "description": "Optional body for POST /auth/logout.\n\n§6 specifies a bodyless logout, so this stays optional and that call keeps\nworking unchanged. Supplying the refresh token narrows the revocation to\nthat one device's token family instead of signing the user out everywhere."
      },
...
      "OtpRequestRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^\\+[1-9]\\d{7,14}$",
            "title": "Phone",
            "examples": [
              "+237600000001"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "OtpRequestRequest"
      },
      "OtpRequestResponse": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid",
            "title": "Challenge Id"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "resend_after_s": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Resend After S"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challenge_id",
          "expires_at",
          "resend_after_s"
        ],
        "title": "OtpRequestResponse",
        "description": "Identical shape for known and unknown numbers. Do not add a field that\nvaries with registration state."
      },
      "OtpVerifyRequest": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid",
            "title": "Challenge Id"
          },
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "pattern": "^\\d{4,8}$",
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challenge_id",
          "code"
        ],
        "title": "OtpVerifyRequest"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 16,
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "RefreshResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "RefreshResponse",
        "description": "Refresh rotates. The token presented here is dead after this call, and\npresenting it again revokes the whole family (Phase 1 reuse detection)."
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "examples": [
              "fr",
              "en"
            ]
          },
          "phone_e164": {
            "type": "string",
            "title": "Phone E164",
            "description": "The caller's own number. Never populated for a counterparty (I3).",
            "examples": [
              "+237600000001"
            ]
          },
          "accessibility": {
            "$ref": "#/components/schemas/AccessibilityProfile"
          },
          "outstanding_xaf": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Outstanding Xaf",
            "description": "Unsettled cancellation debt. Booking is refused with OUTSTANDING_BALANCE while this is above zero.",
            "default": 0
          },
          "driver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "display_name",
          "role",
          "status",
          "locale",
          "phone_e164",
          "accessibility",
          "created_at"
        ],
        "title": "User"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "passenger",
          "driver",
          "admin"
        ],
        "title": "UserRole"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "restricted",
          "suspended"
        ],
        "title": "UserStatus"
      },
...
    }
  }
}